remove os helpers and refocus code
Implementing the OS helpers correctly requires an enormous increase in
code complexity.  A line-based document is needed, and likely a fully
separate context.

This commit admits that there are two problems at play - first the
matter of building helpers and compiling recipes, second a syntax for
writing shell scripts.  Perhaps it was inevitable.